Add authorization flow and secure routes

+++ controllers/feed.js
@@ -5,6 +5,7 @@ const { v4: uuidv4 } = require('uuid');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -48,21 +49,32 @@ exports.postPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     imageUrl: imageUrl,
     content: content,
-    creator: {
-      name: 'Adam',
-    },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      console.log(result);
+      // console.log(result);
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post)
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created!',
-        post: result,
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        }
       });
     })
     .catch((err) => {
@@ -71,7 +83,7 @@ exports.postPost = (req, res, next) => {
       }
       next(err);
     });
-};
+}
 
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
@@ -118,6 +130,12 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      console.log(post);
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -146,12 +164,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // is user logged?
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
+    .then(() => {
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
     .then(result => {
-      console.log(result);
+      // console.log(result);
       res.status(200).json({ message: 'Post deleted!' });
     })
     .catch(err => {
